Refactor notification widget
