+++ lib/ui/home/status_bar_widget.dart
@@ -97,17 +97,21 @@ class _StatusBarWidgetState extends State<StatusBarWidget> {
             ? HeaderErrorWidget(error: _syncError)
             : const SizedBox.shrink(),
         UserRemoteFlagService.instance.shouldShowRecoveryVerification()
-            ? NotificationWarningWidget(
-                warningIcon: Icons.error_outline,
-                actionIcon: Icons.arrow_forward,
-                text: "Confirm your recovery key",
-                onTap: () async => {
-                  await routeToPage(
-                    context,
-                    const VerifyRecoveryPage(),
-                    forceCustomPageRoute: true,
-                  )
-                },
+            ? Padding(
+                padding:
+                    const EdgeInsets.symmetric(horizontal: 16.0, vertical: 12),
+                child: NotificationWidget(
+                  startIcon: Icons.error_outline,
+                  actionIcon: Icons.arrow_forward,
+                  text: "Confirm your recovery key",
+                  onTap: () async => {
+                    await routeToPage(
+                      context,
+                      const VerifyRecoveryPage(),
+                      forceCustomPageRoute: true,
+                    )
+                  },
+                ),
               )
             : const SizedBox.shrink()
       ],
